refactor(analyzer): isolate comb loop SSA evaluation
